# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://admin.lsp-mpsdm.com",
            "description": "Production API Server"
        },
        {
            "url": "https://admin-staging.lsp-mpsdm.com",
            "description": "Staging API Server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development API Server"
        }
    ],
    "paths": {
...
        "/api/apl02": {
            "get": {
                "tags": [
                    "APL02"
                ],
                "summary": "Asesi APL02 Index Data",
                "operationId": "App\\Http\\Controllers\\Api\\Apl02Controller::index",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "APL02"
                ],
                "summary": "Update APL02 Unit Kompetensi Element",
                "operationId": "App\\Http\\Controllers\\Api\\Apl02Controller::updateElement",
                "requestBody": {
                    "description": "Update Unit Kompetensi Element",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "element_id": {
                                        "description": "ID Unit Kompetensi Element",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "type": {
                                        "description": "Informasikan apakah data yang di input itu baru atau perubahan, jika update maka harus input media_id, jika baru tidak perlu.! Input Type: 'new' or 'update'.",
                                        "type": "string",
                                        "example": "new"
                                    },
                                    "media_id": {
                                        "description": "ID Media digunakan jika type update!",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "description": {
                                        "description": "Deskripsi File Upload, Digunakan untuk type: new/update",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "value": {
                                        "description": "File upload. Ekstensi: JPG/JPEG/PNG/PDF",
                                        "type": "file",
                                        "example": "cara_penerapan_xxx.pdf"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/apl02/{id}": {
            "get": {
                "tags": [
                    "APL02"
                ],
                "summary": "APL02 Detail By ID",
                "operationId": "App\\Http\\Controllers\\Api\\Apl02Controller::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sertifikasi id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "APL02"
                ],
                "summary": "Delete APL02 Unit Kompetensi Element Media",
                "operationId": "App\\Http\\Controllers\\Api\\Apl02Controller::deleteElement",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Media ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Sign In",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::login",
                "requestBody": {
                    "description": "Login Credential",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "account email",
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "description": "account password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "pass1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register Account",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::register",
                "requestBody": {
                    "description": "Register Account",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Login Function.",
                                        "type": "string",
                                        "format": "email",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "description": "Login Function.",
                                        "type": "string",
                                        "format": "password",
                                        "example": "pass1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "apiKey",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}